Feature refactoring2

+++ frontend/src/modules/Mbarara/components/ProductList/Product/Product.tsx
@@ -4,12 +4,18 @@ import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
 import Rating from '@mui/material/Rating';
 import FavoriteIcon from '@mui/icons-material/Favorite';
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { Box } from '@mui/material';
 import productStyle from './productStyle';
 import Quantity from '../../../../../shared/Buttons/Quantity';
 import { useAppDispatch, useAppSelector } from '../../../../../hooks/hook';
-import { deleteOrderItem, orderItemsSelector, setOrderItems, updateOrderItem } from '../../../../../redux';
+import {
+  deleteOrderItem,
+  orderItemsSelector,
+  setOrderItems,
+  updateOrderItem,
+} from '../../../../../redux';
+import { consts } from '../../../../../consts/consts';
 
 interface ProductArgs {
   id: number;
@@ -19,85 +25,89 @@ interface ProductArgs {
   discount: number;
   image: string;
   isFavorite: boolean;
-};
+}
 
-const Product = ({ id, name, price, rating, isFavorite, discount, image }: ProductArgs) => {
+const Product = ({
+  id,
+  name,
+  price,
+  rating,
+  isFavorite,
+  discount,
+  image,
+}: ProductArgs) => {
   const dispatch = useAppDispatch();
   const orderItems = useAppSelector(orderItemsSelector);
 
   const handleCount = (id: number) => {
-    for (let obj of orderItems) {
+    for (const obj of orderItems) {
       if (obj.productId === id) {
         return obj.quantity;
       }
     }
-  }
+  };
   const count = handleCount(id);
 
   const [value, setValue] = useState<number | null>(rating);
   const [quantity, setQuantity] = useState<number>(0 | count);
-  const [isFavoriteP, setIsFavorite] = useState<boolean | null>(isFavorite || false);
+  const [isFavoriteP, setIsFavorite] = useState<boolean | null>(
+    isFavorite || false
+  );
 
   const heartStyle = {
     heart: {
       color: isFavoriteP ? 'red' : 'gray',
       height: '17px',
-    }
+    },
   };
 
   const handleSetFavorite = () => {
     if (!isFavoriteP) {
       setIsFavorite(true);
-    }
-    else {
+    } else {
       setIsFavorite(false);
     }
-  }
+  };
+
+  const addOrderItemsObj = {
+    productId: id,
+    orderedPrice: price - (price * discount) / 100,
+    quantity: quantity + 1,
+  };
+
+  const deleteOrderItemsObj = {
+    productId: id,
+    orderedPrice: price - (price * discount) / 100,
+    quantity: quantity - 1,
+  };
 
   const handleIncreaseQuantity = useCallback(() => {
     setQuantity(quantity + 1);
     if (orderItems.length === 0) {
-      dispatch(setOrderItems({
-        productId: id,
-        orderedPrice: price - ((price * discount) / 100),
-        quantity: quantity + 1
-      }))
+      dispatch(setOrderItems(addOrderItemsObj));
     } else {
-      for (let obj of orderItems) {
+      for (const obj of orderItems) {
         if (obj.productId === id) {
-          dispatch(updateOrderItem({
-            productId: id,
-            orderedPrice: price - ((price * discount) / 100),
-            quantity: quantity + 1
-          }))
+          dispatch(updateOrderItem(addOrderItemsObj));
         }
         if (obj.productId !== id) {
-          dispatch(deleteOrderItem(id))
-          dispatch(setOrderItems({
-            productId: id,
-            orderedPrice: price - ((price * discount) / 100),
-            quantity: quantity + 1
-          }))
+          dispatch(deleteOrderItem(id));
+          dispatch(setOrderItems(addOrderItemsObj));
         }
       }
     }
-  }, [quantity])
+  }, [quantity]);
+
   const handleDecreaseQuantity = () => {
     setQuantity(quantity - 1);
-    for (let obj of orderItems) {
-      if (obj.productId === id) {
-        if (obj.quantity > 1) {
-          dispatch(updateOrderItem({
-            productId: id,
-            orderedPrice: price - ((price * discount) / 100),
-            quantity: quantity - 1
-          }))
-        } else {
-          dispatch(deleteOrderItem(id))
-        }
+    for (const obj of orderItems) {
+      if (obj.productId === id && obj.quantity > 1) {
+        dispatch(updateOrderItem(deleteOrderItemsObj));
+      } else {
+        dispatch(deleteOrderItem(id));
       }
     }
-  }
+  };
 
   return (
     <Box sx={productStyle.MainBox}>
@@ -106,16 +116,18 @@ const Product = ({ id, name, price, rating, isFavorite, discount, image }: Produ
           style={{ backgroundSize: 'cover' }}
           component='img'
           height='110'
-          image={`https://shopimages54643.s3.eu-west-2.amazonaws.com/${image}`}
+          image={`${consts.bucketUrl}${image}`}
           alt={name}
         />
-        {
-          discount ?
-            <div style={productStyle.DiscoundBlock as React.CSSProperties}>
-              <button disabled style={productStyle.DiscoundBtn}>-{discount}% off</button>
-            </div>
-            : ''
-        }
+        {discount ? (
+          <div style={productStyle.DiscoundBlock as React.CSSProperties}>
+            <button disabled style={productStyle.DiscoundBtn}>
+              -{discount}% off
+            </button>
+          </div>
+        ) : (
+          ''
+        )}
         <CardContent>
           <Typography component='div'>
             {name.length > 15 ? `${name.substr(0, 15)}...` : name}
@@ -126,25 +138,25 @@ const Product = ({ id, name, price, rating, isFavorite, discount, image }: Produ
                 <Rating name='read-only' size='small' value={value} readOnly />
               </div>
               <div style={productStyle.HeartBlock as React.CSSProperties}>
-                <FavoriteIcon className='heart' onClick={handleSetFavorite} style={heartStyle.heart} />
+                <FavoriteIcon
+                  className='heart'
+                  onClick={handleSetFavorite}
+                  style={heartStyle.heart}
+                />
               </div>
             </div>
           </Typography>
           <div style={productStyle.PriceBlock as React.CSSProperties}>
-            {
-              discount ?
-                <>
-                  <div style={productStyle.PriceDiscText}>
-                    {price - ((price * discount) / 100)}$
-                  </div>
-                  <div style={productStyle.PriceText}>
-                    {price}$
-                  </div>
-                </>
-                : <div style={productStyle.PriceDiscText}>
-                  {price}$
+            {discount ? (
+              <>
+                <div style={productStyle.PriceDiscText}>
+                  {price - (price * discount) / 100}$
                 </div>
-            }
+                <div style={productStyle.PriceText}>{price}$</div>
+              </>
+            ) : (
+              <div style={productStyle.PriceDiscText}>{price}$</div>
+            )}
             <Quantity
               quantity={quantity}
               handleIncreaseQuantity={handleIncreaseQuantity}
@@ -155,6 +167,6 @@ const Product = ({ id, name, price, rating, isFavorite, discount, image }: Produ
       </Card>
     </Box>
   );
-}
+};
 
 export default Product;
